fix dtype issues

+++ model.py
@@ -69,7 +69,7 @@ class Transformer(nn.Module):
         if targets is not None:
             return linear_cross_entropy(
                 x.to(self.dtype),
-                self.classifier.weight,
+                self.classifier.weight.to(self.dtype),
                 targets,
                 accum_e_fp32=True,
                 accum_c_fp32=True,
